# Use a linked list of background tasks to perform