--- a/ports/atmel-samd/background.c
+++ b/ports/atmel-samd/background.c
@@ -39,59 +39,21 @@
 #include "shared-module/displayio/__init__.h"
 #endif
 
-volatile uint64_t last_finished_tick = 0;
-
-bool stack_ok_so_far = true;
-
-static bool running_background_tasks = false;
-
 #ifdef MONITOR_BACKGROUND_TASKS
 // PB03 is physical pin "SCL" on the Metro M4 express
 // so you can't use this code AND an i2c peripheral
 // at the same time unless you change this
-STATIC void start_background_task(void) {
+void port_start_background_task(void) {
     REG_PORT_DIRSET1 = (1<<3);
     REG_PORT_OUTSET1 = (1<<3);
 }
 
-STATIC void finish_background_task(void) {
+void port_finish_background_task(void) {
     REG_PORT_OUTCLR1 = (1<<3);
 }
 #else
-STATIC void start_background_task(void) {}
-STATIC void finish_background_task(void) {}
+void port_start_background_task(void) {}
+void port_finish_background_task(void) {}
 #endif
 
-void background_tasks_reset(void) {
-    running_background_tasks = false;
-}
-
-void run_background_tasks(void) {
-    // Don't call ourselves recursively.
-    if (running_background_tasks) {
-        return;
-    }
-
-    start_background_task();
-
-    assert_heap_ok();
-    running_background_tasks = true;
-
-    #if CIRCUITPY_DISPLAYIO
-    displayio_background();
-    #endif
-
-    #if CIRCUITPY_NETWORK
-    network_module_background();
-    #endif
-    filesystem_background();
-    running_background_tasks = false;
-    assert_heap_ok();
-
-    last_finished_tick = port_get_raw_ticks(NULL);
-    finish_background_task();
-}
-
-bool background_tasks_ok(void) {
-    return port_get_raw_ticks(NULL) - last_finished_tick < 1024;
-}
+void port_background_task(void) {}