--- a/ports/cxd56/background.h
+++ b/ports/cxd56/background.h
@@ -27,7 +27,4 @@
 #ifndef MICROPY_INCLUDED_CXD56_BACKGROUND_H
 #define MICROPY_INCLUDED_CXD56_BACKGROUND_H
 
-void background_tasks_reset(void);
-void run_background_tasks(void);
-
 #endif  // MICROPY_INCLUDED_CXD56_BACKGROUND_H