--- a/ports/esp32s2/background.c
+++ b/ports/esp32s2/background.c
@@ -35,27 +35,12 @@
 #include "shared-module/displayio/__init__.h"
 #endif
 
-static bool running_background_tasks = false;
 
-void background_tasks_reset(void) {
-    running_background_tasks = false;
-}
-
-void run_background_tasks(void) {
-    // Don't call ourselves recursively.
-    if (running_background_tasks) {
-        return;
-    }
-
+void port_background_task(void) {
     // Zero delay in case FreeRTOS wants to switch to something else.
     vTaskDelay(0);
-    running_background_tasks = true;
-    filesystem_background();
-
-    #if CIRCUITPY_DISPLAYIO
-    displayio_background();
-    #endif
-    running_background_tasks = false;
-
-    assert_heap_ok();
 }
+
+void port_start_background_task(void) {}
+
+void port_finish_background_task(void) {}