--- a/ports/litex/background.h
+++ b/ports/litex/background.h
@@ -27,9 +27,4 @@
 #ifndef MICROPY_INCLUDED_LITEX_BACKGROUND_H
 #define MICROPY_INCLUDED_LITEX_BACKGROUND_H
 
-#include <stdbool.h>
-
-void background_tasks_reset(void);
-void run_background_tasks(void);
-
 #endif  // MICROPY_INCLUDED_LITEX_BACKGROUND_H