--- a/ports/mimxrt10xx/background.c
+++ b/ports/mimxrt10xx/background.c
@@ -24,58 +24,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-#include "background.h"
 
-//#include "audio_dma.h"
-#include "supervisor/filesystem.h"
-#include "supervisor/shared/tick.h"
-#include "supervisor/usb.h"
+#include "supervisor/port.h"
 
-#include "py/runtime.h"
-#include "shared-module/network/__init__.h"
-#include "supervisor/linker.h"
-#include "supervisor/shared/stack.h"
-
-#ifdef CIRCUITPY_DISPLAYIO
-#include "shared-module/displayio/__init__.h"
-#endif
-
-volatile uint64_t last_finished_tick = 0;
-
-bool stack_ok_so_far = true;
-
-static bool running_background_tasks = false;
-
-void background_tasks_reset(void) {
-    running_background_tasks = false;
-}
-
-void PLACE_IN_ITCM(run_background_tasks)(void) {
-    // Don't call ourselves recursively.
-    if (running_background_tasks) {
-        return;
-    }
-    assert_heap_ok();
-    running_background_tasks = true;
-
+void port_background_task(void) {
     #if CIRCUITPY_AUDIOIO || CIRCUITPY_AUDIOBUSIO
     audio_dma_background();
     #endif
-    #if CIRCUITPY_DISPLAYIO
-    displayio_background();
-    #endif
-
-    #if CIRCUITPY_NETWORK
-    network_module_background();
-    #endif
-    filesystem_background();
-    usb_background();
-    running_background_tasks = false;
-    assert_heap_ok();
-
-    last_finished_tick = supervisor_ticks_ms64();
 }
-
-bool background_tasks_ok(void) {
-    return supervisor_ticks_ms64() - last_finished_tick < 1000;
-}
+void port_start_background_task(void) {}
+void port_finish_background_task(void) {}