--- a/ports/nrf/background.c
+++ b/ports/nrf/background.c
@@ -46,35 +46,14 @@
 #include "common-hal/_bleio/bonding.h"
 #endif
 
-static bool running_background_tasks = false;
+void port_start_background_task(void) {}
+void port_finish_background_task(void) {}
 
-void background_tasks_reset(void) {
-    running_background_tasks = false;
-}
-
-void run_background_tasks(void) {
-    // Don't call ourselves recursively.
-    if (running_background_tasks) {
-        return;
-    }
-    running_background_tasks = true;
-    filesystem_background();
+void port_background_task(void) {
 #if CIRCUITPY_AUDIOPWMIO
     audiopwmout_background();
 #endif
 #if CIRCUITPY_AUDIOBUSIO
     i2s_background();
 #endif
-
-#if CIRCUITPY_BLEIO
-    supervisor_bluetooth_background();
-    bonding_background();
-#endif
-
-    #if CIRCUITPY_DISPLAYIO
-    displayio_background();
-    #endif
-    running_background_tasks = false;
-
-    assert_heap_ok();
 }