The "lower power" branch removed a previous optimization that allowed the background tasks to run just once every tick.  Restore this, for things that run on the basis of time; and where possible move things to happening based on interrupts.

The foundation for this is a new linked list of background callbacks.  A background callback takes a function pointer and an optional object pointer.  Usually, such callbacks are registered from interrupts.  For instance, on atmel sam, the audio DMA interrupt is used to schedule a fill of the new sample data, rather than polling a flag every 1ms.

Current state of this PR:
 * Builds on a range of boards
 * transitions all boards back to running background tasks just once per ms, though only when ticks are enabled
 * transitions sam d20/d5x/e5x to using interrupts for audio and usb

Testing performed: on pygamer and grand central m4, did general tire kicking of related code:
 * audio code
 * usb
 * filesystem
 * eink display
 * lcd display
 * gamepad

Benchmarks:  On Grand Central M4, I ran the following code
```
import time
t0 = time.monotonic_ns()
k = 0
for i in range(1000):
    for j in range(100):
        k += j
t1 = time.monotonic_ns()
print(k)
print((t1-t0) / 1e9)
```
The regression from 5.3.0 to 6.0.0.alpha.1 is almost entirely fixed.  However, the difference between the original timing of 1.02s and the new timing of 1.04s seems "real" and is not within the usual variability of execution time. (that the above loop takes about 1s on samd51 is a total coincidence, I didn't choose the numbers specially or anything!)

| version | time |
|:-------------| -----:|
| 5.3.0   |  1.02s |
| 6.0.0.a1 |  1.54s |
| this PR | 1.04s |

What's next: 
 * test & benchmark more boards
 * get a green from Actions
 * possible strategic conversion of more stuff to interrupt-driven