@jepler 2020-06-25
Discussion with @tannewt 
- go ahead with this approach
- make sure that the case of background callbacks becoming invalid during GC heap reset 

Which modules to do next
- the rest of audio (it's "the hardest")
- displayio second
- anything that has open issues about it not working in main