@jepler 2020-07-09
The initial comment has been heavily revised to reflect the current status of this PR.  Things are progressing nicely.  SAM D5x/E5x is probably done, though as always there may be problems that only become obvious after testing, such as in the next alpha release.