@jepler 2020-07-09
The performance gained back on the CLUE is much smaller... but it has not been given specific attention yet.

| version | time |
|:-------------| -----:|
| 5.3.0   |  2.65s |
| 6.0.0.a1 |  4.66s |
| this PR | 4.112s |