@jepler 2020-07-10
With the latest changes performance on nRF is a bit better than the 5.3.0 baseline (now testing on Particle Xenon).

| version | time |
|:-------------| -----:|
| 5.3.0   |  2.24s |
| 6.0.0.a1 |  3.04s |
| this PR | 2.22ss |

I suspect the 5.3.0 performance of the Xenon is better than the CLUE because the latter has a displayio display enabled.  However, I didn't test the theory.