@jepler 2020-07-17
@tannewt thanks for reminding me about that case.  By making the callback list a GC root, this should not be possible.  Please re-review.