@tannewt 2020-07-17
@jepler are all callbacks allocated on the heap or is it a mix? If it's a mix then the collect won't work because it'll stop at the first pointer off the heap I believe.

Looks like uchip is 4 bytes too big now unfortunately.